% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGate.R
\name{ggGate}
\alias{ggGate}
\title{Create gates from ggplot2 object}
\usage{
ggGate(p, write_data_to = "df_new", write_gate_to = "df_gate")
}
\arguments{
\item{p}{ggplot2 object.}

\item{write_data_to}{String. Sets the variable name of the \code{data.frame} that stores the annotated data underlying the ggplot2 object.}

\item{write_gate_to}{String. Sets the variable name of the \code{data.frame} that stores the gating coordinates.}
}
\description{
Create gates by clicking on ggplot2.
A single click adds coordinates to the active polygon gate.
A double click draws the active polygon gate, and annotates the raw ggplot2 data with new columns.
}
